refactor: add trainer profile
