+++ src/pages/profile/trainer/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Profile: React.FC = () => {
+  return <></>;
+};
+
+export default Profile;
