+++ src/pages/profile/user/components/cards/trainer_card.tsx
@@ -1,7 +1,9 @@
 import { Card, Row } from "antd";
 import Text from "antd/lib/typography/Text";
-import Translations from "../../../../localization/translations";
-import ButtonContact, { ContactType } from "../../../../shared/button_contact";
+import Translations from "../../../../../localization/translations";
+import ButtonContact, {
+  ContactType,
+} from "../../../../../shared/button_contact";
 import React from "react";
 import { useTranslation } from "react-i18next";
