+++ src/pages/profile/user/profile.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { Provider } from "react-redux";
-import { store } from "../../redux/store";
+import { store } from "../../../redux/store";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import React from "react";
 import Profile from "./index";
